- refactor folder structure
