+++ src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {QueryClient, QueryClientProvider} from 'react-query';
-import Router from './router';
+import Router from './routes';
 
 const queryClient = new QueryClient();
